Standard App: change io.c/io.h
